+++ karma-test-shim.js
@@ -43,6 +43,7 @@ System.config({
   // Extend usual application package list with test folder
   packages: {
     rxjs: { defaultExtension: 'js' },
+    'rxjs/operators': { defaultExtension: 'js', main: 'index' },
     '': { defaultExtension: 'js' },
     src: {
         defaultExtension: 'js',